(PATCH) hash buffer size & block_count type
/*
 *                                      hashes time index
 * buffer size > nonce + 64*2 + 10 + 1  =  n + 139
 * Let's consider a max nonce length of 11 digits
*/

unsigned int block_count(Blockchain* blockchain);

+++ src/level-1/1-dcc-c/database/Blockchain.h
@@ -28,4 +28,4 @@ Blockchain* hash_block(Blockchain* b);
 
 void clear_blockchain(blockchainPtr* blockchain);
 
-size_t block_count(Blockchain* blockchain);
+unsigned int block_count(Blockchain* blockchain);
